// dops/src/enum/cache-key.enum.ts
export enum CacheKey {
  COUNTRY = 'COUNTRY',
  PROVINCE = 'PROVINCE',
  PERMIT_TYPE = 'PERMIT_TYPE',
  VEHICLE_TYPE = 'VEHICLE_TYPE',
  POWER_UNIT_TYPE = 'POWER_UNIT_TYPE',
  TRAILER_TYPE = 'TRAILER_TYPE',
  EMAIL_TEMPLATE_PROFILE_REGISTRATION = 'EMAIL_TEMPLATE_PROFILE_REGISTRATION',
  EMAIL_TEMPLATE_ISSUE_PERMIT = 'EMAIL_TEMPLATE_ISSUE_PERMIT',
  EMAIL_TEMPLATE_PAYMENT_RECEIPT = 'EMAIL_TEMPLATE_PAYMENT_RECEIPT',
  EMAIL_TEMPLATE_COMPANY_SUSPEND = 'EMAIL_TEMPLATE_COMPANY_SUSPEND',
  EMAIL_TEMPLATE_COMPANY_UNSUSPEND = 'EMAIL_TEMPLATE_COMPANY_UNSUSPEND',
  EMAIL_TEMPLATE_APPLICATION_APPROVED = 'EMAIL_TEMPLATE_APPLICATION_APPROVED',
  EMAIL_TEMPLATE_APPLICATION_REJECTED = 'EMAIL_TEMPLATE_APPLICATION_REJECTED',
  EMAIL_TEMPLATE_ORBC_STYLE = 'EMAIL_TEMPLATE_ORBC_STYLE',
  CHES_ACCESS_TOKEN = 'CHES_ACCESS_TOKEN',
  CDOGS_ACCESS_TOKEN = 'CDOGS_ACCESS_TOKEN',
  DOCUMENT_TEMPLATE = 'DOCUMENT_TEMPLATE',
  PAYMENT_AND_REFUND_DETAILED_REPORT = 'PAYMENT_AND_REFUND_DETAILED_REPORT',
  PAYMENT_AND_REFUND_SUMMARY_REPORT = 'PAYMENT_AND_REFUND_SUMMARY_REPORT',
  FEATURE_FLAG_TYPE = 'FEATURE_FLAG_TYPE',
  IMG_HEADER_LOGO = 'IMG_HEADER_LOGO',
  IMG_FOOTER_LOGO = 'IMG_FOOTER_LOGO',
  IMG_DARK_MODE_HEADER_LOGO = 'IMG_DARK_MODE_HEADER_LOGO',
  IMG_DARK_MODE_MED_HEADER_LOGO = 'IMG_DARK_MODE_MED_HEADER_LOGO',
  IMG_DARK_MODE_FOOTER_LOGO = 'IMG_DARK_MODE_FOOTER_LOGO',
  IMG_DARK_MODE_MED_FOOTER_LOGO = 'IMG_DARK_MODE_MED_FOOTER_LOGO',
  IMG_WHITE_HEADER_LOGO = 'IMG_WHITE_HEADER_LOGO',
  IMG_WHITE_MED_HEADER_LOGO = 'IMG_WHITE_MED_HEADER_LOGO',
  IMG_WHITE_FOOTER_LOGO = 'IMG_WHITE_FOOTER_LOGO',
  IMG_WHITE_MED_FOOTER_LOGO = 'IMG_WHITE_MED_FOOTER_LOGO',
  IMG_WHITE_HEADER_ORBC_LOGO = 'IMG_WHITE_HEADER_ORBC_LOGO',
  IMG_WHITE_MED_HEADER_ORBC_LOGO = 'IMG_WHITE_MED_HEADER_ORBC_LOGO',
}